requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
transformers
torch --index-url https://download.pytorch.org/whl/cpu
PyMuPDF
numpy
pydantic